+++ libs/eavmlib/src/esp.erl
@@ -38,6 +38,9 @@
     sleep_enable_ulp_wakeup/0,
     deep_sleep/0,
     deep_sleep/1,
+    light_sleep/0,
+    sleep_enable_gpio_wakeup/0,
+    sleep_enable_timer_wakeup/1,
     mount/4,
     umount/1,
     nvs_fetch_binary/2,
@@ -275,6 +278,36 @@ sleep_enable_ulp_wakeup() ->
 %%-----------------------------------------------------------------------------
 -spec deep_sleep() -> no_return().
 deep_sleep() ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
+%% @doc Put the esp32 into light sleep.
+%% This function returns ok on success after the sleep.
+%% Program is NOT restarted and wake up reason can be inspected to determine how
+%% the esp32 was woken up.
+%% @end
+%%-----------------------------------------------------------------------------
+-spec light_sleep() -> ok | error.
+light_sleep() ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
+%% @doc Enable waking up from light sleep using a configured GPIO
+%% @returns `ok | error'
+%% @end
+%%-----------------------------------------------------------------------------
+-spec sleep_enable_gpio_wakeup() -> ok | error.
+sleep_enable_gpio_wakeup() ->
+    erlang:nif_error(undefined).
+
+%%-----------------------------------------------------------------------------
+%% @doc Enable waking up from light sleep using a timer
+%% @param time before wakeup in microseconds
+%% @returns `ok | error'
+%% @end
+%%-----------------------------------------------------------------------------
+-spec sleep_enable_timer_wakeup(SleepUS :: non_neg_integer()) -> ok | error.
+sleep_enable_timer_wakeup(_SleepUS) ->
     erlang:nif_error(undefined).
 
 %%-----------------------------------------------------------------------------