+++ CHANGELOG.md
@@ -111,6 +111,8 @@ instead `badarg`.
 
 - Added `lists:keysort/2`
 - Added `lists:merge/2,3`
+- [ESP32] Added support to light sleep (`esp:light_sleep/0`) and to GPIO and timer wakeup from
+light sleep.
 
 ### Fixed